% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Export.R
\name{RE_calculateall}
\alias{RE_calculateall}
\title{Do all non-custom calculations at once}
\usage{
RE_calculateall(list)
}
\arguments{
\item{list}{List with raw Rock-Eval data}
}
\value{
Converted list with all basic metrics
}
\description{
Combines all main functions in one go. Not recommended for first time use.
}
