% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thermo_Analysis.R
\name{RE_Ccombined}
\alias{RE_Ccombined}
\title{Determine the total C released at each time step}
\usage{
RE_Ccombined(list)
}
\arguments{
\item{list}{List with converted Rock-Eval data}
}
\value{
Combined C flux at each time step
}
\description{
Combines the amount of C released of the different thermograms to calculate the flux of PC and RC over time.
}
\section{How to use}{

This function calculates the amount of released C at every time step (instead of integrating it over an entire period).
The input needs to be data converted to the right units by using the function RE_convert().
}

\section{Parameter info}{

\itemize{
\item POC, pyrolysis organic C (mg C/ g / s)
\item PIC, pyrolysis inorganic C (mg C/ g / s)
\item PC,  pyrolysis total C (mg C/ g / s)
\item ROC, residual organic C (mg C/ g / s)
\item RIC, residual inorganic C (mg C/ g / s)
\item RC,  residual total inorganic C (mg C/ g / s)
}
}

