% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Editing.R
\name{RE_convert}
\alias{RE_convert}
\title{Convert Rock-Eval signal to right units}
\usage{
RE_convert(list)
}
\arguments{
\item{list}{List with raw Rock-Eval data from RE_read}
}
\value{
Input list with Rock-Eval thermograms in proper units (mg/g)
}
\description{
This function derives the elemental amount using calibration, baseline and sample weight.
}
\section{How to use}{

It is important to convert the thermogram data to the right units, before using most other functions.
This function performs the conversion using the callibration values and baseline.
The input data should be the list of raw data files which is read by the function RE_read(dataloc).
}

