% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Editing.R
\name{RE_cumtraparea}
\alias{RE_cumtraparea}
\title{Cumulative trapezoidal integration}
\usage{
RE_cumtraparea(x, y)
}
\arguments{
\item{x}{Independent variable for integral (e.g. time)}

\item{y}{Dependent variable of same length as x}
}
\value{
Trapezoidal area under time series data
}
\description{
This function does a step wise trapezoidal integration.
}
