% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Export.R
\name{RE_extracttable}
\alias{RE_extracttable}
\title{Extract all time independent Rock-Eval metrics}
\usage{
RE_extracttable(list)
}
\arguments{
\item{list}{List with converted and extended Rock-Eval data}
}
\value{
Simplest possible data type containing requested data
}
\description{
Creates a table with all calculated metrics added to the data.
}
\section{How to use}{

When all parameters of interest have been calculated, this function will convert the list of samples into a single data.frame for further analysis.
}

