% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Viewing.R
\name{RE_plot}
\alias{RE_plot}
\title{Plot Rock-Eval thermograms}
\usage{
RE_plot(datalist, analysis, oven = "both", thermogram = "all")
}
\arguments{
\item{datalist}{List containing data of one or more Rock-Eval analyses}

\item{analysis}{Name of file containing data of one Rock-Eval analysis}

\item{oven}{Choice of pyrolysis or oxidation oven}

\item{thermogram}{Choice of selected thermogram}
}
\value{
A plot of thermogram(s)
}
\description{
This function previews Rock-Eval thermograms and cursors. Input contains a single Rock-Eval analysis.
}
\section{How to use}{

This function allows for basic plotting of the raw data and cursors + baselines.
First data needs to be read into the right format by using the RE_read() function.
By default all the thermograms are shown, but the oven and type of thermogram are manually adjustable.
}

