% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thermo_Analysis.R
\name{RE_Tpercentiles}
\alias{RE_Tpercentiles}
\title{Determine relative T percentiles}
\usage{
RE_Tpercentiles(list, vector = c(25, 50, 75))
}
\arguments{
\item{list}{List with converted Rock-Eval data and zone areas}

\item{vector}{Vector with the percentiles to be calculated}
}
\value{
Temperature percentiles
}
\description{
Calculates the temperature percentiles of the thermograms
}
\section{How to use}{

Before using this function the data need to be converted to the right units by using the function RE_convert() and the Rock-Eval zones need to be calculated with the function RE_surfaces().
From these surfaces the temperature at which the given percentiles have been reached is derived (e.g., S3CO2-T50 is the temperature at which 50\% of the S3CO2 has been pyrolysed).
Default percentiles are 25,50,75 but any vector containing numeric values can be used for this calculation.
}

