% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Editing.R
\name{RE_traparea}
\alias{RE_traparea}
\title{Trapezoidal integration}
\usage{
RE_traparea(x, y)
}
\arguments{
\item{x}{Independent variable for integral (e.g. time)}

\item{y}{Dependent variable of same length as x}
}
\value{
Trapezoidal area under time series data
}
\description{
This function does a simplified trapezoidal integration.
}
