% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RockEvalR.R
\docType{package}
\name{RockEvalR}
\alias{RockEvalR-package}
\alias{RockEvalR}
\title{RockEvalR: Convert and use Rock-Eval raw data.}
\description{
This package allows the user to calculate basic Rock-Eval parameters and do additional operations.
The package reads .B00 files into an efficient R-format for further use.
Raw data have preferably been accessed by Geoworks such that calibration information was applied.
}
\author{
\strong{Maintainer}: Maarten van Hoef \email{maartenvanhoef@gmail.com}

}
\keyword{internal}
