% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thermo_Analysis.R
\name{RE_Ccombined}
\alias{RE_Ccombined}
\title{Determine the total C released at each time step}
\usage{
RE_Ccombined(list)
}
\arguments{
\item{list}{List with converted Rock-Eval data}
}
\value{
Combined C flux at each time step
}
\description{
Combines the amount of C released of the different thermograms to calculate the flux of PC and RC over time.
}
