% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Editing.R
\name{RE_convert}
\alias{RE_convert}
\title{Convert Rock-Eval signal to right units}
\usage{
RE_convert(list)
}
\arguments{
\item{list}{List with raw Rock-Eval data from RE_read}
}
\value{
Input list with Rock-Eval thermograms in proper units (mg/g)
}
\description{
This function derives the elemental amount using calibration, baseline and sample weight.
}
