% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom_Analysis.R
\name{RE_customarea}
\alias{RE_customarea}
\title{Calculate custom Rock-Eval surfaces}
\usage{
RE_customarea(list, oven, thermogram, temperatures)
}
\arguments{
\item{list}{List with converted Rock-Eval data from RE_read}

\item{oven}{Pyrolysis or Oxidation oven}

\item{thermogram}{Must be one of the columns of the oven data, e.g., CO2}

\item{vector}{Vector containing temperature intervals (excluding 0 and max)}
}
\value{
Input list with included areas of Rock-Eval zones
}
\description{
This function calculates the areas between a set of temperatures. Input should be converted already.
The areas between all temperatures is given plus the area between t0 and the first temperature and the area after the highest temperature is reached.
All output parameters have the same units as their fully integrated counterparts.
}
