% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom_Analysis.R
\name{RE_customarea}
\alias{RE_customarea}
\title{Calculate custom Rock-Eval surfaces}
\usage{
RE_customarea(list, oven, thermogram, temperatures)
}
\arguments{
\item{list}{List with converted Rock-Eval data from RE_read}

\item{oven}{Pyrolysis or Oxidation oven}

\item{thermogram}{Must be one of the columns of the oven data, e.g., CO2}

\item{temperatures}{Vector containing temperature intervals (excluding 0 and max)}
}
\value{
Input list with included areas of Rock-Eval zones
}
\description{
This function calculates the areas between a set of temperatures. Input should be converted already.
The areas between all temperatures is given plus the area between t0 and the first temperature and the area after the highest temperature is reached.
All output parameters have the same units as their fully integrated counterparts.
}
\section{How to use}{

It is important that first the data have been converted to the right units (by calibration values) and that the baseline is set to 0.
Converting the raw measurements to proper units and baselined values can be done with the RE_convert() function.
This function integrates any thermograms. Integration happens between the temperatures that are given as a vector argument in the function.
The lowest temperature and highest temperatures should not be included in the argument vector.
So vector c(300,450,500) will result in the integration: T < 300, T 300-450, T 450-500, and T > 500.
}

