% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Analysis.R
\name{RE_metrics}
\alias{RE_metrics}
\title{Calculate Rock-Eval metrics}
\usage{
RE_metrics(list)
}
\arguments{
\item{list}{List with converted Rock-Eval data and zone areas}
}
\value{
Computed Rock-Eval metrics
}
\description{
This function calculates main Rock-Eval metrics from zone areas.
}
\section{Parameter info}{

\itemize{
\item Tpeak, RE7 temperature at S2 peak (C)
\item POC, pyrolysis organic C (\%)
\item ROC, residual organic C (\%)
\item TOC, total organic C (\%)
\item PIC, pyrolysis inorganic C (\%)
\item RIC, residual inorganic C (\%)
\item TIC, total inorganic C (\%)
\item HI, hydrogen index (mg CH/ g TOC)
\item OI_CO, CO oxygen index (mg CO/ g TOC)
\item OI_CO2, main oxygen index (mg CO2/ g TOC)
\item OI_combi, combined oxygen index/ OIRE6 (mg O/ g TOC)
}
}

