% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Import.R
\name{RE_read}
\alias{RE_read}
\title{Load Rock-Eval data files}
\usage{
RE_read(dataloc)
}
\arguments{
\item{dataloc}{Path to the input file}
}
\value{
A list of the files in dataloc.
}
\description{
This function loads all Rock-Eval raw data files into a list.
}
