% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thermo_Analysis.R
\name{RE_SebagIR}
\alias{RE_SebagIR}
\title{Sebag (2016) areas and I/R indices}
\usage{
RE_SebagIR(list)
}
\arguments{
\item{list}{List with converted Rock-Eval data and zone areas}
}
\value{
Computed indices from Sebag (2016)
}
\description{
Calculates the areas A1, A2, A3, A4 plus the I and R indices.
}
\references{
Sebag, et al. (2016). Dynamics of soil organic matter based on new Rock-Eval indices. Geoderma, 284, 185-203.
}
