% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sulfur_Functions.R
\name{RE_Smetrics}
\alias{RE_Smetrics}
\title{Calculate Rock-Eval Sulphur metrics}
\usage{
RE_Smetrics(list, org.conversion.a = -1.9, org.conversion.b = 843)
}
\arguments{
\item{list}{List with converted Rock-Eval data and zone areas}

\item{org.conversion.a}{Slope of pyrolysis to total organic S empirical conversion}

\item{org.conversion.b}{Intercept of pyrolysis to total organic S empirical conversion}
}
\value{
Computed Rock-Eval metrics
}
\description{
This function calculates main Rock-Eval Sulphur metrics from zone areas.
}
\section{How to use}{

It is necessary that the RE Sulphur surfaces have been integrated first.
This can be done with the RE_Ssurfaces() function.
Additionally, for the sulphur index (SI) the total C has to be determined with the RE_surfaces() function.
From these areas the basic Rock-Eval metrics are then computed.
The determination of most sulphur-species depends on empirical correction of matrix effects.
Based on type of sample (e.g., Kerogen type), and data availability different empirical adjustments may lead to better results.

In this case total organic S is calculated from a relationship with the Tmax_CH during pyrolysis:

(Pyrolysed org S) / (Total org S) = (org.conversion.a) x (Tmax_CH) + (org.conversion.b)

(Total inorg S) = (Total S) - (Total org S)
}

\section{Parameter info}{

\itemize{
\item Tpeak_S, RE7 temperature at organic SO2 peak during pyrolysis (C)
\item POS, pyrolysis organic S (\%)
\item PIS, pyrolysis inorganic S (\%)
\item RTS, total oxidised S (\%)
\item TS, total measured S (\%)
\item PyOS.emp, emperically determined ratio of pyrolysed organic S (\%)
\item ROS.cor, empirically corrected residual organic S (\%)
\item TOS.cor, empirically corrected total organic S (\%)
\item RIS.cor, empirically corrected residual inorganic S (\%)
\item TIS.cor, empirically corrected total inorganic S (\%)
\item SI, sulphur index (mg total org. S/ g TOC)
\item SI.pyr, labile sulphur index (mg pyrolysed org. S/ g TOC)
}
}

\references{
Cohen-Sadon, et al. (2022). A new empirical approach for rapid quantification of organic and pyritic sulphur in sedimentary rocks using the Rock-Eval 7S. Organic Geochemistry, 166, 104350.

Aboussou (2018). New Rock-Eval method for Pyritic and Organic Sulphur quantification: Application to study Organic matter preservation in Jurassic sediments. Earth Sciences, Sorbonne Université.

Cohen-Sadon, et al. (2025). Tmax-S: A new proxy for the role of sulphur on sedimentary organic matter preservation and thermal maturation. Journal of Analytical and Applied Pyrolysis 190 - 107115.
}
\seealso{
\code{\link[=RE_surfaces]{RE_surfaces()}}
}
