% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sulfur_Functions.R
\name{RE_Ssurfaces}
\alias{RE_Ssurfaces}
\title{Calculate the Rock-Eval Sulphur surfaces}
\usage{
RE_Ssurfaces(list, time.include = FALSE)
}
\arguments{
\item{list}{List with converted Rock-Eval data}

\item{time.include}{Include the curves as individual time series TRUE/FALSE}
}
\value{
Input list with included areas of Rock-Eval zones
}
\description{
This function calculates the areas of the different Rock-Eval zones for Sulphur. Input should be converted already. As an option the zones can also be included as a time series.
}
\section{How to use}{

It is important that first the data have been converted to the right units (by calibration values) and that the baseline is set to 0.
Converting the raw measurements to proper units and baselined values can be done with the RE_convert() function.
This function integrates the different thermograms for S. Integration happens between the cursors that were determined automatically or manually adjusted.
Here the areas and nomenclature of Cohen-Sadon et al. (2022) are used instead of those in Geoworks. However, the cut-off temperatures are not fixed an based on the cursors and local minima.
}

\section{Parameter info}{

\itemize{
\item S1S, Labile pyrolysable organic S (mg S/g)
\item S2S, Pyrolysable organic S (mg S/g)
\item S3S, Pyrolysable inorganic S (mg S/g)
\item S4S, Oxidised S (mg S/g)
\item S5S, Residual oxidised S (mg S/g)
}
}

\section{Cut-off info}{

\itemize{
\item S1S, pyrolysed before start of the heating ramp
\item S2S, pyrolysed from start of heating until minimum between 475 and 550 C
\item S3S, pyrolysed after minimum between 475 and 550 C
\item S4S, oxidised below minimum between 550 and 650 C
\item S5S, oxidised above minimum between 550 and 650 C
}
}

\references{
Cohen-Sadon, et al. (2022). A new empirical approach for rapid quantification of organic and pyritic sulphur in sedimentary rocks using the Rock-Eval 7S. Organic Geochemistry, 166, 104350.

Aboussou (2018). New Rock-Eval method for Pyritic and Organic Sulphur quantification: Application to study Organic matter preservation in Jurassic sediments. Earth Sciences, Sorbonne Université.
}
