% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Editing.R
\name{RE_surfaces}
\alias{RE_surfaces}
\title{Calculate the Rock-Eval surfaces}
\usage{
RE_surfaces(list)
}
\arguments{
\item{list}{List with converted Rock-Eval data from RE_read}
}
\value{
Input list with included areas of Rock-Eval zones
}
\description{
This function calculates the areas of the different Rock-Eval zones. Input should be converted already.
}
