% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Analysis.R
\name{RE_surfaces}
\alias{RE_surfaces}
\title{Calculate the Rock-Eval surfaces}
\usage{
RE_surfaces(list, time.inlude = FALSE)
}
\arguments{
\item{list}{List with converted Rock-Eval data}

\item{time.include}{Include the curves as individual time series TRUE/FALSE}
}
\value{
Input list with included areas of Rock-Eval zones
}
\description{
This function calculates the areas of the different Rock-Eval zones. Input should be converted already. As an option the zones can also be included as a timeseries.
}
