% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thermo_Analysis.R
\name{RE_Tpercentiles}
\alias{RE_Tpercentiles}
\title{Determine relative T percentiles}
\usage{
RE_Tpercentiles(list, vector = c(25, 50, 75))
}
\arguments{
\item{list}{List with converted Rock-Eval data and zone areas}

\item{vector}{Vector with the percentiles to be calculated}
}
\value{
Temperature percentiles
}
\description{
Calculates the temperature percentiles of the thermograms
}
